Splint and cppcheck cleanup.  All regression tests pass.

+++ driver_evermore.c
@@ -183,8 +183,8 @@ gps_mask_t evermore_parse(struct gps_device_t * session, unsigned char *buf,
 
     switch (type) {
     case 0x02:			/* Navigation Data Output */
-	session->context->gps_week = getleuw(buf2, 2);
-	session->context->gps_tow = getleul(buf2, 4) * 0.01;
+	session->context->gps_week = (unsigned short)getleuw(buf2, 2);
+	session->context->gps_tow = (double)getleul(buf2, 4) * 0.01;
 	/*@ ignore @*//*@ splint is confused @ */
 	session->newdata.time =
 	    gpstime_to_unix(session->context->gps_week,
@@ -223,8 +223,8 @@ gps_mask_t evermore_parse(struct gps_device_t * session, unsigned char *buf,
 	return mask | CLEAR_IS | REPORT_IS;
 
     case 0x04:			/* DOP Data Output */
-	session->context->gps_week = getleuw(buf2, 2);
-	session->context->gps_tow = getleul(buf2, 4) * 0.01;
+	session->context->gps_week = (unsigned short)getleuw(buf2, 2);
+	session->context->gps_tow = (double)getleul(buf2, 4) * 0.01;
 	/*@ ignore @*//*@ splint is confused @ */
 	session->newdata.time =
 	    gpstime_to_unix(session->context->gps_week,
@@ -267,8 +267,8 @@ gps_mask_t evermore_parse(struct gps_device_t * session, unsigned char *buf,
 	return mask;
 
     case 0x06:			/* Channel Status Output */
-	session->context->gps_week = getleuw(buf2, 2);
-	session->context->gps_tow = getleul(buf2, 4) * 0.01;
+	session->context->gps_week = (unsigned short)getleuw(buf2, 2);
+	session->context->gps_tow = (double)getleul(buf2, 4) * 0.01;
 	/*@ ignore @*//*@ splint is confused @ */
 	session->gpsdata.skyview_time =
 	    gpstime_to_unix(session->context->gps_week,
@@ -327,8 +327,8 @@ gps_mask_t evermore_parse(struct gps_device_t * session, unsigned char *buf,
     case 0x08:			/* Measurement Data Output */
 	/* clock offset is a manufacturer diagnostic */
 	/* (int)getleuw(buf2, 8);  clock offset, 29000..29850 ?? */
-	session->context->gps_week = getleuw(buf2, 2);
-	session->context->gps_tow = getleul(buf2, 4) * 0.01;
+	session->context->gps_week = (unsigned short)getleuw(buf2, 2);
+	session->context->gps_tow = (double)getleul(buf2, 4) * 0.01;
 	/*@ ignore @*//*@ splint is confused @ */
 	session->newdata.time =
 	    gpstime_to_unix(session->context->gps_week,
